updated custom webfont icons tree sample
