+++ src/frontend-samples/tree-samples/custom-webfont-icons-tree/CustomWebfontIconsTree.scss
@@ -1,0 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+* See LICENSE.md in the project root for license terms and full copyright notice.
+*--------------------------------------------------------------------------------------------*/
+.tree {
+  height: 100%;
+}
